+++ packages/storybook/stories/IncomingCallNotification/InboundCalling.stories.mdx
@@ -4,7 +4,7 @@ import InboundCallingSource from '!!raw-loader!./snippets/IncomingCallStackProps
 
 <Meta
   id="inboundcalling"
-  title="Concepts/InboundCalling"
+  title="Concepts/Inbound Calling"
   parameters={{ previewTabs: { canvas: { disable: true, hidden: true } } }}
 />
