+++ packages/storybook/stories/IncomingCallNotification/snippets/IncomingCallStackPropsExample.snippet.tsx
